feat(checkout): CHECKOUT-6269 PaymentMethodProps
