+++ .eslintrc.json
@@ -24,6 +24,12 @@
               }
             ]
           }
+        ],
+        "@typescript-eslint/no-unused-vars": [
+          "warn",
+          {
+            "varsIgnorePattern": "^_"
+          }
         ]
       }
     },
