+++ packages/core/src/app/core/paymentIntegration/index.ts
@@ -1,4 +1,1 @@
-export { default as PaymentMethodProps } from './PaymentMethodProps';
-export { default as ResolvableComponent } from './ResolvableComponent';
-export { default as toResolvableComponent } from './toResolvableComponent';
 export { default as resolvePaymentMethod, PaymentMethodResolveId } from './resolvePaymentMethod';
