+++ packages/core/src/app/payment/Payment.tsx
@@ -1,4 +1,4 @@
-import { PaymentMethodId, PaymentMethodProviderType } from '@bigcommerce/checkout-js/payment-integration';
+import { PaymentMethodId, PaymentMethodProviderType, PaymentFormValues } from '@bigcommerce/checkout-js/payment-integration';
 import { CartChangedError, CheckoutSelectors, CheckoutSettings, OrderRequestBody, PaymentMethod } from '@bigcommerce/checkout-sdk';
 import { memoizeOne } from '@bigcommerce/memoize';
 import { compact, find, isEmpty, noop } from 'lodash';
@@ -16,7 +16,7 @@ import mapSubmitOrderErrorMessage, { mapSubmitOrderErrorTitle } from './mapSubmi
 import mapToOrderRequestBody from './mapToOrderRequestBody';
 import { getUniquePaymentMethodId } from './paymentMethod';
 import PaymentContext from './PaymentContext';
-import PaymentForm, { PaymentFormValues } from './PaymentForm';
+import PaymentForm from './PaymentForm';
 
 export interface PaymentProps {
     isEmbedded?: boolean;
