+++ packages/core/src/app/payment/createPaymentFormService.ts
@@ -1,10 +1,10 @@
+import { PaymentFormService } from '@bigcommerce/checkout-js/payment-integration';
 import { FormikContext } from 'formik';
 
 import { FormContextType } from '../ui/form';
 
 import { PaymentContextProps } from './PaymentContext';
 import { PaymentFormValues } from './PaymentForm';
-import PaymentFormService from './PaymentFormService';
 
 export default function createPaymentFormService(
     formikContext: FormikContext<PaymentFormValues>,
