+++ packages/core/src/app/payment/creditCard/index.ts
@@ -1,5 +1,5 @@
 export { default as configureCardValidator } from './configureCardValidator';
-export { default as CreditCardFieldset, CreditCardFieldsetProps, CreditCardFieldsetValues } from './CreditCardFieldset';
+export { default as CreditCardFieldset, CreditCardFieldsetProps } from './CreditCardFieldset';
 export { default as CreditCardIcon } from './CreditCardIcon';
 export { default as CreditCardIconList } from './CreditCardIconList';
 export { default as CreditCardCodeField } from './CreditCardCodeField';
