+++ packages/core/src/app/payment/hostedCreditCard/index.ts
@@ -1,3 +1,2 @@
-export { HostedCreditCardValidationValues } from './HostedCreditCardValidation';
 export { default as withHostedCreditCardFieldset, WithInjectedHostedCreditCardFieldsetProps } from './withHostedCreditCardFieldset';
-export { default as HostedCreditCardFieldset, HostedCreditCardFieldsetValues } from './HostedCreditCardFieldset';
+export { default as HostedCreditCardFieldset } from './HostedCreditCardFieldset';
