+++ packages/core/src/app/payment/index.ts
@@ -1,3 +1,2 @@
 export { PaymentProps } from './Payment';
-export { default as PaymentFormService } from './PaymentFormService';
 export { default as getPreselectedPayment } from './getPreselectedPayment';
