+++ packages/core/src/app/payment/paymentMethod/PaymentMethodTitle.spec.tsx
@@ -1,4 +1,4 @@
-import { PaymentMethodId, PaymentMethodType } from '@bigcommerce/checkout-js/payment-integration';
+import { PaymentMethodId, PaymentMethodType, PaymentFormValues } from '@bigcommerce/checkout-js/payment-integration';
 import { createCheckoutService, CheckoutService, StoreConfig } from '@bigcommerce/checkout-sdk';
 import { mount } from 'enzyme';
 import { Formik } from 'formik';
@@ -10,7 +10,6 @@ import { getStoreConfig } from '../../config/config.mock';
 import { createLocaleContext, LocaleContext, LocaleContextType } from '../../locale';
 import { CreditCardIconList } from '../creditCard';
 import { getPaymentMethod } from '../payment-methods.mock';
-import { PaymentFormValues } from '../PaymentForm';
 
 import getPaymentMethodName from './getPaymentMethodName';
 import PaymentMethodTitle, { PaymentMethodTitleProps } from './PaymentMethodTitle';
