+++ packages/core/src/app/payment/storedInstrument/getInstrumentValidationSchema.ts
@@ -1,3 +1,4 @@
+import { CardInstrumentFieldsetValues } from '@bigcommerce/checkout-js/payment-integration';
 import { LanguageService } from '@bigcommerce/checkout-sdk';
 import { memoize } from '@bigcommerce/memoize';
 import { cvv, number } from 'card-validator';
@@ -5,7 +6,6 @@ import creditCardType from 'credit-card-type';
 import { object, string, ObjectSchema, StringSchema } from 'yup';
 
 import mapFromInstrumentCardType from './mapFromInstrumentCardType';
-import { CardInstrumentFieldsetValues } from './CardInstrumentFieldset';
 
 export interface InstrumentValidationSchemaOptions {
     instrumentBrand: string;
