+++ packages/core/src/app/payment/storedInstrument/index.ts
@@ -5,7 +5,7 @@ export { default as isInstrumentCardCodeRequiredSelector } from './isInstrumentC
 export { default as isInstrumentCardNumberRequired } from './isInstrumentCardNumberRequired';
 export { default as isInstrumentCardNumberRequiredSelector } from './isInstrumentCardNumberRequiredSelector';
 export { default as AccountInstrumentFieldset } from './AccountInstrumentFieldset';
-export { default as CardInstrumentFieldset, CardInstrumentFieldsetValues } from './CardInstrumentFieldset';
+export { default as CardInstrumentFieldset } from './CardInstrumentFieldset';
 export { default as CreditCardValidation } from './CreditCardValidation';
 export { default as isCardInstrument } from './isCardInstrument';
 export { default as isAccountInstrument } from './isAccountInstrument';
