+++ packages/core/src/app/paymentIntegration/anotherSample/AnotherSamplePaymentMethod.tsx
@@ -1,6 +1,5 @@
+import { toResolvableComponent, PaymentMethodProps } from '@bigcommerce/checkout-js/payment-integration';
 import React, { FunctionComponent } from 'react';
-
-import { toResolvableComponent, PaymentMethodProps } from '../../core/paymentIntegration';
 
 const AnotherSamplePaymentMethod: FunctionComponent<PaymentMethodProps> = () => <></>;
