+++ packages/payment-integration/project.json
@@ -7,7 +7,7 @@
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["packages/payment-integration/**/*.ts"]
+        "lintFilePatterns": ["packages/payment-integration/**/*.{ts,tsx}"]
       }
     },
     "test": {
