+++ packages/payment-integration/src/PaymentFormService.ts
@@ -1,7 +1,7 @@
 import { PaymentMethod } from '@bigcommerce/checkout-sdk';
 import { ObjectSchema } from 'yup';
 
-import { PaymentFormValues } from './PaymentForm';
+import PaymentFormValues from './PaymentFormValues';
 
 export default interface PaymentFormService {
     disableSubmit(method: PaymentMethod, disabled?: boolean): void;
