+++ packages/payment-integration/src/PaymentMethodId.ts
@@ -1,4 +1,4 @@
-export enum PaymentMethodId {
+enum PaymentMethodId {
     Adyen = 'adyen',
     AdyenV2 = 'adyenv2',
     AdyenV2GooglePay = 'googlepayadyenv2',
@@ -59,3 +59,5 @@ export enum PaymentMethodId {
     WorldpayAccess = 'worldpayaccess',
     Zip = 'zip',
 }
+
+export default PaymentMethodId;
