+++ packages/payment-integration/src/PaymentMethodProviderType.ts
@@ -1,6 +1,8 @@
-export enum PaymentMethodProviderType {
+enum PaymentMethodProviderType {
     Api = 'PAYMENT_TYPE_API',
     Hosted = 'PAYMENT_TYPE_HOSTED',
     Offline = 'PAYMENT_TYPE_OFFLINE',
     PPSDK = 'PAYMENT_TYPE_SDK',
-};
+}
+
+export default PaymentMethodProviderType;
