+++ packages/payment-integration/src/ResolvableComponent.ts
