+++ packages/payment-integration/src/index.ts
@@ -1,4 +1,20 @@
-export { HostedPaymentMethodProps } from './HostedPaymentMethod';
-export { PaymentMethodId } from './PaymentMethodId';
-export { PaymentMethodProviderType } from './PaymentMethodProviderType';
-export { PaymentMethodType } from './PaymentMethodType';
+export { default as HostedPaymentMethodProps } from './HostedPaymentMethod';
+export { default as PaymentFormService } from './PaymentFormService';
+export { default as PaymentMethodId } from './PaymentMethodId';
+export { default as PaymentMethodProps } from './PaymentMethodProps';
+export { default as PaymentMethodProviderType } from './PaymentMethodProviderType';
+export { default as PaymentMethodType } from './PaymentMethodType';
+export { default as ResolvableComponent } from './ResolvableComponent';
+export { default as toResolvableComponent } from './toResolvableComponent';
+export { default as isResolvableComponent } from './isResolvableComponent';
+export {
+    default as PaymentFormValues,
+    CardInstrumentFieldsetValues,
+    CreditCardFieldsetValues,
+    DocumentOnlyCustomFormFieldsetValues,
+    FawryCustomFormFieldsetValues,
+    HostedCreditCardFieldsetValues,
+    HostedCreditCardValidationValues,
+    IdealCustomFormFieldsetValues,
+    SepaCustomFormFieldsetValues,
+} from './PaymentFormValues';
