+++ packages/payment-integration/tsconfig.spec.json
@@ -4,7 +4,7 @@
     "types": ["jest", "node"]
   },
   "include": [
-    "**/*.test.ts",
-    "**/*.spec.ts"
+    "**/*.spec.ts",
+    "**/*.spec.tsx"
   ]
 }
